requirements.txt

numpy==1.26.4
pandas==2.2.2
scikit-learn==1.5.2

fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2

sentence-transformers==3.0.1
faiss-cpu==1.12.0


tqdm==4.66.4
joblib==1.4.2

sentence-transformers==3.0.1
torch==2.5.1
torchvision==0.20.1
torchaudio==2.5.1
transformers>=4.40.0
